Modularize: Change ES bundle to output individual files instead of a single bundle.
This makes tree shaking much better.

This is a breaking change, but I can fix that by adding a forwarder.
