+++ README.md
@@ -1,4 +1,5 @@
 # @hebcal/core
+
 Hebcal is a perpetual Jewish Calendar. This library converts between
 Hebrew and Gregorian dates, and generates lists of Jewish holidays for
 any year (past, present or future).  Shabbat and holiday candle
@@ -20,11 +21,23 @@ Many users of this library will utilize the [HebrewCalendar](https://hebcal.gith
 and [HDate](https://hebcal.github.io/api/hdate/classes/HDate.html) interfaces.
 
 ## Installation
+
 ```bash
-$ npm install @hebcal/core
+npm install @hebcal/core
 ```
 
+## Usage
+
+This package exports 3 categories of output:
+
+- `dist/bundle.min.js`: This includes all dependencies in a single JS file.  It's quite large (174KB minified), and should only be used if you aren't using a bundler.
+- `dist/index.cjs`: This includes all dependencies as a single CommonJS module, including polyfills.  Avoid where possible.
+- `dist/es/`: This directory contains ES modules for each source file in the package.  If you're using Rollup, Vite, or another bundler, you should import from here to reduce output sizes.
+  - These files do not include polyfills; if you want to support older browsers, use <https://github.com/rollup/plugins/tree/master/packages/babel#babelhelpers>
+  - To make tree-shaking work better, import from the most specific file possible; avoid importing from `index` or `hebcal`.  Some APIs may still need to be moved to specific files.
+
 ## Synopsis
+
 ```javascript
 import {HebrewCalendar, HDate, Location, Event} from '@hebcal/core';
 
@@ -44,4 +57,5 @@ for (const ev of events) {
   console.log(date.toLocaleDateString(), ev.render('en'), hd.toString());
 }
 ```
+
 ## [API Documentation](https://hebcal.github.io/api/core/index.html)
