+++ size-demo/getHoliday.ts
@@ -1,2 +1,2 @@
-import {HDate, HebrewCalendar} from '../dist/index';
+import {HDate, HebrewCalendar} from '../dist/es/index';
 console.log(HebrewCalendar.getHolidaysOnDate(new HDate(1, 1, 5757)));
