+++ size-demo/sedra.ts
@@ -1,3 +1,3 @@
-import {getSedra} from '../dist/index';
+import {getSedra} from '../dist/es/sedra';
 const sedra = getSedra(5757, false);
 console.log(sedra.get(56));
