+++ size-demo/sizes.md
@@ -1,8 +1,8 @@
 # Bundle sizes
 
-- `getHoliday.mjs`: 524 kB
-- `getHoliday.min.mjs`: 208 kB
-- `parshiyot.mjs`: 524 kB
-- `parshiyot.min.mjs`: 208 kB
-- `sedra.mjs`: 216 kB
-- `sedra.min.mjs`: 102 kB
+- `getHoliday.mjs`: 525 kB
+- `getHoliday.min.mjs`: 209 kB
+- `parshiyot.mjs`: 525 kB
+- `parshiyot.min.mjs`: 209 kB
+- `sedra.mjs`: 89.4 kB
+- `sedra.min.mjs`: 33.3 kB
